AI/requirements.txt
Flask
pdfplumber
google-genai
pandas
xgboost
scikit-learn
joblib
requests
matplotlib
python-dotenv
sqlalchemy
pymysql